Refactor user store into modular utilities
