+++ backend/src/utils/userStore/helpers.ts
@@ -1,0 +1,10 @@
+export function slugify(name: string): string {
+  return name
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/^-+|-+$/g, "");
+}
+
+export function maskFromLastFour(lastFour: string): string {
+  return `**** **** **** ${lastFour}`;
+}
