+++ backend/src/utils/userStore/index.ts
@@ -1,0 +1,46 @@
+export { isInternalOrg } from "./config";
+export { generatePlainApiKey, revealStoredKey } from "./apiKeys";
+export { createPasswordHash, verifyPassword } from "./passwords";
+export { getIdentityStore } from "./persistence";
+export {
+  getOrganizations,
+  getOrganization,
+  createOrganizationWithOwner,
+  topUpOrganization,
+  recordUsage,
+} from "./organizations";
+export {
+  getUsersForOrganization,
+  getUser,
+  getUserByEmail,
+  createUserAccount,
+  updateUserLastLogin,
+  attachUserToOrganization,
+  createOrUpdateOrgUser,
+} from "./users";
+export {
+  addKeySet,
+  removeKeySet,
+  rotateApiKey,
+  findOrgByApiKey,
+} from "./keyManagement";
+export {
+  toSafeKey,
+  toSafeKeySet,
+  toSafeOrganization,
+  toSafeUser,
+  maskKey,
+} from "./safeEntities";
+export {
+  summarizeUsageEntries,
+  summarizeTopUps,
+  getPlatformOverview,
+} from "./summaries";
+
+export type { SafeEntityOptions } from "./safeEntities";
+export type {
+  UsageTotals,
+  TopUpTotals,
+  PlatformOrganizationSummary,
+  PlatformOverview,
+} from "./summaries";
