+++ backend/src/utils/userStore/passwords.ts
@@ -1,0 +1,16 @@
+import crypto from "crypto";
+
+export function createPasswordHash(password: string): string {
+  const salt = crypto.randomBytes(16);
+  const hash = crypto.scryptSync(password, salt, 64);
+  return `${salt.toString("hex")}:${hash.toString("hex")}`;
+}
+
+export function verifyPassword(password: string, storedHash: string): boolean {
+  const [saltHex, hashHex] = storedHash.split(":");
+  if (!saltHex || !hashHex) return false;
+  const salt = Buffer.from(saltHex, "hex");
+  const hash = Buffer.from(hashHex, "hex");
+  const derived = crypto.scryptSync(password, salt, hash.length);
+  return crypto.timingSafeEqual(hash, derived);
+}
