+++ backend/src/utils/userStore/time.ts
@@ -1,0 +1,3 @@
+export function now(): string {
+  return new Date().toISOString();
+}
